work in progress

+++ lib/bloc/fly_bloc/fly_event.dart
@@ -10,77 +10,77 @@ abstract class FlyEvent extends Equatable {
 class FlyCheckConnectionEvent extends FlyEvent {}
 
 class GyroXUpdated extends FlyEvent {
-  final int value;
+  final double value;
   GyroXUpdated(this.value);
 }
 
 class GyroYUpdated extends FlyEvent {
-  final int value;
+  final double value;
   GyroYUpdated(this.value);
 }
 
 class GyroZUpdated extends FlyEvent {
-  final int value;
+  final double value;
   GyroZUpdated(this.value);
 }
 
 class MagnetometerXUpdated extends FlyEvent {
-  final int value;
+  final double value;
   MagnetometerXUpdated(this.value);
 }
 
 class MagnetometerYUpdated extends FlyEvent {
-  final int value;
+  final double value;
   MagnetometerYUpdated(this.value);
 }
 
 class MagnetometerZUpdated extends FlyEvent {
-  final int value;
+  final double value;
   MagnetometerZUpdated(this.value);
 }
 
 class AccelerometerXUpdated extends FlyEvent {
-  final int value;
+  final double value;
   AccelerometerXUpdated(this.value);
 }
 
 class AccelerometerYUpdated extends FlyEvent {
-  final int value;
+  final double value;
   AccelerometerYUpdated(this.value);
 }
 
 class AccelerometerZUpdated extends FlyEvent {
-  final int value;
+  final double value;
   AccelerometerZUpdated(this.value);
 }
 
 class BarometerUpdated extends FlyEvent {
-  final int value;
+  final double value;
   BarometerUpdated(this.value);
 }
 
 class Motor1SpeedUpdated extends FlyEvent {
-  final int value;
+  final double value;
   Motor1SpeedUpdated(this.value);
 }
 
 class Motor2SpeedUpdated extends FlyEvent {
-  final int value;
+  final double value;
   Motor2SpeedUpdated(this.value);
 }
 
 class BatterySocUpdated extends FlyEvent {
-  final int value;
+  final double value;
   BatterySocUpdated(this.value);
 }
 
 class BatteryVolUpdated extends FlyEvent {
-  final int value;
+  final double value;
   BatteryVolUpdated(this.value);
 }
 
 class SignalUpdated extends FlyEvent {
-  final int value;
+  final double value;
   SignalUpdated(this.value);
 }
 
@@ -112,17 +112,17 @@ class SendYoke extends FlyEvent {
 }
 
 class YawUpdated extends FlyEvent {
-  final int value;
+  final double value;
   YawUpdated(this.value);
 }
 
 class RollUpdated extends FlyEvent {
-  final int value;
+  final double value;
   RollUpdated(this.value);
 }
 
 class PitchUpdated extends FlyEvent {
-  final int value;
+  final double value;
   PitchUpdated(this.value);
 }
